# docs/openapi.yaml
openapi: 3.1.0
info:
  title: LeakProof AI API
  version: 0.1.0
  description: API contract for the Gemini-backed revenue audit workspace.
paths:
...
  /api/findings/{id}/status:
    patch:
      summary: Update finding status
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [approved, dismissed, needs_review, customer_ready, recovered, not_recoverable]
                note:
                  type: string
      responses:
        '200':
          description: Updated
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer